pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.7.9",
    "chromadb>=1.1.0",
    "faiss-cpu>=1.12.0",
    "flagembedding>=1.3.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-ollama>=0.3.8",
    "langchain-openai>=0.3.33",
    "langchain-text-splitters>=0.3.11",
    "langgraph>=1.0.0a3",
    "langgraph-cli[inmem]>=0.4.2",
    "loguru>=0.7.3",
    "pymilvus>=2.6.2",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
    "python-socks>=2.7.2",
    "sentence-transformers>=5.1.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.2",
    "trl>=0.23.0",
]
